% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doPDPA.R
\name{doPDPA}
\alias{doPDPA}
\title{High-level function for segmentation by pruned dynamic programming algorithm (pDPA)}
\usage{
doPDPA(Y, K, ...)
}
\arguments{
\item{Y}{A numeric vector the signal to be segmented}

\item{K}{The number of change points to find}

\item{\dots}{Further parameters to be passed to Fpsn}
}
\description{
High-level function for segmentation by pruned dynamic programming algorithm (pDPA)
}
\examples{
## load known real copy number regions
affyDat <- loadCnRegionData(dataSet="GSE29172", tumorFraction=1)

## generate a synthetic CN profile
K <- 10
len <- 1e4
sim <- getCopyNumberDataByResampling(len, K, minLength=100, regData=affyDat)
datS <- sim$profile

## run pruned DPA segmentation
res <- doPDPA(datS[["c"]], K=2*K)
getTpFp(res$bkp, sim$bkp, tol=5, relax = -1)   ## true and false positives
plotSeg(datS, breakpoints=list(sim$bkp, res$bkp))
}
\references{
Rigaill, G. (2015). A pruned dynamic programming algorithm to recover the best segmentations with 1 to K_max change-points. Journal de la Soci\\u00e9t\\u00e9 Fran\\u00e7aise de Statistique, 156(4), 180-205.
}
\author{
Morgane Pierre-Jean, Guillem Rigaill and Pierre Neuvial
}
