% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruned_DP_C_access.R
\name{Fpsn}
\alias{Fpsn}
\title{Call the pruned DPA algorithm, use functional pruning and segment neighborhood.}
\usage{
Fpsn(x, Kmax, mini = min(x), maxi = max(x))
}
\arguments{
\item{x}{A vector of double : the signal to be segmented}

\item{Kmax}{Max number of segments}

\item{mini}{Min value for the mean parameter of the segment}

\item{maxi}{Max value for the mean parameter of the segment}
}
\value{
A list with a vector containing the position of the change-points
}
\description{
Call the pruned DPA algorithm, use functional pruning and segment neighborhood.
}
\details{
The pruned DPA is implemented for the L2-loss function
}
\author{
Guillem Rigaill
}
