% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUnivJ.R
\name{getUnivJ}
\alias{getUnivJ}
\title{Get the contribution of one dimension to the RSE.}
\usage{
getUnivJ(y, candCP)
}
\arguments{
\item{y}{Input vector of signals for the considered dimension}

\item{candCP}{A vector of candidate change points}
}
\description{
Get the contribution of one dimension to the Residual Squared Error (RSE)
}
\details{
This function is used internally by \code{\link{pruneByDP}}.
}
\seealso{
\code{\link{pruneByDP}}
}
\author{
Morgane Pierre-Jean and Pierre Neuvial
}
\keyword{internal}
