\name{jointseg-package}
\alias{jointseg-package}
\alias{jointseg}
\docType{package}
\title{Joint segmentation of multivariate (copy number) signals}
}
\description{This package implements functions to quickly segment multivariate signals into piecewise constant profiles and a framework to generate realistic copy-number profiles.  A typical application is the joint segmentation of total DNA copy numbers and allelic ratios obtained from Single Nucleotide Polymorphism (SNP) microarrays in cancer studies.}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{jointseg}
\packageIndices{jointseg}
}
\author{Morgane Pierre-Jean and Pierre Neuvial}

Maintainer: \packageMaintainer{jointseg}
}

\keyword{ package }


