% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapPositionsBack.R
\name{mapPositionsBack}
\alias{mapPositionsBack}
\title{mapPositionsBack}
\usage{
mapPositionsBack(pos)
}
\arguments{
\item{pos}{A sorted list of position (or indices)}
}
\description{
Map breakpoint positions back to original space
}
\author{
Morgane Pierre-Jean and Pierre Neuvial
}
