% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.R
\name{prof}
\alias{prof}
\title{profile time and memory usage of a given R expression}
\usage{
prof(expr, doit = TRUE)
}
\arguments{
\item{expr}{An \code{R} expression to be evaluated}

\item{doit}{A boolean variable specifying whether profiling should be
performed or not (intended for internal use).}
}
\description{
profile time and memory usage of a given R expression
}
\details{
Profiling is performed using \code{summaryRprof(memory="both")$by.self}.
}
\note{
Our memory profiling is not satistfactory yet!
}
\seealso{
\code{\link{Rprof}}

\code{\link{summaryRprof}}
}
\author{
Morgane Pierre-Jean and Pierre Neuvial
}
\keyword{internal}
