% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruned_DP_C_access.R
\name{retour_sn}
\alias{retour_sn}
\title{Extract endpoint matrix from DP result}
\usage{
retour_sn(path)
}
\arguments{
\item{path}{the path vector of the "colibri_sn_R_c C" function}
}
\description{
Extract endpoint matrix from DP result
}
