% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruned_DP_C_access.R
\name{Fpsn}
\alias{Fpsn}
\title{Pruned dynamic programming algorithm}
\usage{
Fpsn(x, Kmax, mini = min(x), maxi = max(x))
}
\arguments{
\item{x}{A vector of double : the signal to be segmented}

\item{Kmax}{Max number of segments}

\item{mini}{Min value for the mean parameter of the segment}

\item{maxi}{Max value for the mean parameter of the segment}
}
\value{
A list with a vector containing the position of the change-points
}
\description{
Low-level API for the pruned dynamic programming algorithm (pDPA)
}
\details{
This implementation uses functional pruning and segment neighborhood, and the L2-loss function
}
\references{
Rigaill, G. (2015). A pruned dynamic programming algorithm to recover the best segmentations with 1 to K_max change-points. Journal de la Soci\\u00e9t\\u00e9 Fran\\u00e7aise de Statistique, 156(4), 180-205.
}
\seealso{
\code{\link{doDynamicProgramming}} for a higher-level function
}
\author{
Guillem Rigaill
}
