\name{jointseg-package}
\alias{jointseg-package}
\alias{jointseg}
\docType{package}
\title{Joint segmentation of multivariate (copy number) signals}
\description{This package implements functions to quickly segment multivariate signals into piecewise constant profiles and a framework to generate realistic copy-number profiles.  A typical application is the joint segmentation of total DNA copy numbers and allelic ratios obtained from Single Nucleotide Polymorphism (SNP) microarrays in cancer studies.}
\details{
\tabular{ll}{Package: \tab jointseg\cr
Type: \tab Package\cr
Title: \tab Joint segmentation of multivariate (copy number) signals\cr
Version: \tab 0.6.2\cr
Date: \tab 2014-09-08\cr
Author: \tab Morgane Pierre-Jean and Pierre Neuvial\cr
Maintainer: \tab Morgane Pierre-Jean <morgane.pierrejean@genopole.cnrs.fr>\cr
License: \tab LGPL (>= 2.1)\cr
Depends: \tab R (>= 2.10), acnr (>= 0.2.0)\cr
Imports: \tab matrixStats (>= 0.6.0), DNAcopy\cr
Suggests: \tab cghseg, PSCBS, aroma.light, R.cache, digest, changepoint (>= 1.0.2), knitr, RUnit, BiocGenerics\cr
VignetteBuilder: \tab knitr\cr
biocViews: \tab Software\cr}
}
\author{Morgane Pierre-Jean and Pierre Neuvial}

\keyword{ package }


