\name{mapPositionsBack}
\alias{mapPositionsBack}
\title{mapPositionsBack}
\description{Map breakpoint positions back to original space}
\usage{mapPositionsBack(pos)}\arguments{
\item{pos}{A sorted list of position (or indices)}
}



\author{Morgane Pierre-Jean and Pierre Neuvial}





